renamed libray loader action to loader action
